remove compound commands
