+++ step-generation/src/commandCreators/compound/index.ts
@@ -6,9 +6,6 @@ export { airGapInWell } from './airGapInWell'
 export { blowOutInTrash } from './blowOutInTrash'
 export { blowOutInWasteChute } from './blowOutInWasteChute'
 export { consolidate } from './consolidate'
-export { delayInTrash } from './delayInTrash'
-export { delayInWasteChute } from './delayInWasteChute'
-export { delayInWell } from './delayInWell'
 export { dispenseInTrash } from './dispenseInTrash'
 export { dispenseInWasteChute } from './dispenseInWasteChute'
 export { distribute } from './distribute'
